Add Footer component and refactor NavigationBar; introduce navigation links constant
