+++ src/constant/navigation_links.js
@@ -1,0 +1,18 @@
+import end_points from "../router/end_points";
+
+const linkItems = [
+  {
+    label: "Home",
+    path: end_points.home,
+  },
+  {
+    label: "Explore",
+    path: end_points.explore,
+  },
+  {
+    label: "Account",
+    path: end_points.account,
+  },
+];
+
+export default linkItems;
